Add role management dialog: implement role creation functionality and enhance user interface for role details

+++ CoreAccess.BizLayer/Services/RoleService.cs
@@ -7,10 +7,11 @@ namespace CoreAccess.BizLayer.Services;
 public interface IRoleService
 {
     Task<PagedResult<RoleDto>> SearchRolesAsync(RoleSearchOptions options, CancellationToken cancellationToken = default);
-    Task<RoleDetailDto> GetRoleByIdAsync(string id, CancellationToken cancellationToken = default);
+    Task<RoleDetailDto> GetRoleByIdAsync(string id, bool includeUsers = false, bool includePermissions = false, CancellationToken cancellationToken = default);
     Task<RoleDetailDto> CreateRoleAsync(RoleCreateRequest request, CancellationToken cancellationToken = default);
     Task<RoleDetailDto> UpdateRoleAsync(string userId, RoleUpdateRequest user, CancellationToken cancellationToken = default);
     Task<RoleDetailDto> AddPermissionToRoleAsync(string roleId, string permissionId, CancellationToken cancellationToken = default);
+    Task<RoleDetailDto> RemovePermissionFromRoleAsync(string roleId, string permissionId, CancellationToken cancellationToken = default);
     Task DeleteRoleAsync(string id, CancellationToken cancellationToken = default);
 }
 
@@ -36,11 +37,11 @@ public class RoleService(IRoleRepository roleRepository, IPermissionRepository p
         }
     }
 
-    public async Task<RoleDetailDto> GetRoleByIdAsync(string id, CancellationToken cancellationToken = default)
+    public async Task<RoleDetailDto> GetRoleByIdAsync(string id, bool includeUsers = false, bool includePermissions = false, CancellationToken cancellationToken = default)
     {
         try
         {
-            var role = await roleRepository.SearchRolesAsync(new RoleSearchOptions() { Id = id }, cancellationToken).ContinueWith(t => t.Result.FirstOrDefault() ?? null, cancellationToken);
+            var role = await roleRepository.SearchRolesAsync(new RoleSearchOptions() { Id = id, IncludeUsers = includeUsers, IncludePermissions = includePermissions}, cancellationToken).ContinueWith(t => t.Result.FirstOrDefault() ?? null, cancellationToken);
             if (role == null)
             {
                 return null;
@@ -156,10 +157,69 @@ public class RoleService(IRoleRepository roleRepository, IPermissionRepository p
         }
     }
 
+    public async Task<RoleDetailDto> RemovePermissionFromRoleAsync(string roleId, string permissionId, CancellationToken cancellationToken = default)
+    {
+        try
+        {
+            var role = await roleRepository.SearchRolesAsync(new RoleSearchOptions()
+            {
+                Id = roleId,
+                IncludePermissions = true,
+                Page = 1,
+                PageSize = 1
+            }, cancellationToken).ContinueWith(t => t.Result.FirstOrDefault() ?? null, cancellationToken);
+            
+            if (role == null)
+            {
+                throw new Exception($"Role '{roleId}' not found");
+            }
+
+            var permission = await permissionRepository.SearchPermissionsAsync(new PermissionSearchOptions()
+            {
+                Id = permissionId,
+                Page = 1,
+                PageSize = 1
+            }, cancellationToken).ContinueWith(x => x.Result.FirstOrDefault() ?? null, cancellationToken);
+            
+            if (permission == null)
+            {
+                throw new Exception($"Permission '{permissionId}' not found");
+            }
+
+            if (role.Permissions.Count == 0 || !role.Permissions.Contains(permission))
+            {
+                throw new Exception($"Role {role.Name} does not have permission {permission.Name}");
+            }
+            
+            role.Permissions.Remove(permission);
+            var updatedRole = await roleRepository.InsertOrUpdateRoleAsync(role, cancellationToken);
+
+            return updatedRole.ToDetailDto();
+        }
+        catch (Exception ex)
+        {
+            throw;
+        }
+    }
+
     public async Task DeleteRoleAsync(string id, CancellationToken cancellationToken = default)
     {
         try
         {
+            var role = await roleRepository.SearchRolesAsync(new RoleSearchOptions()
+            {
+                Id = id,
+                IncludeUsers = true,
+                Page = 1,
+                PageSize = 1
+            }, cancellationToken).ContinueWith(t => t.Result.FirstOrDefault() ?? null, cancellationToken);
+            
+            if(role == null)
+                throw new Exception("Role not found");
+            
+            if(role.Users.Count > 0)
+                throw new Exception($"Cannot delete role with assigned users. Please remove all {role.Users.Count} users.");
+            
             await roleRepository.DeleteRoleAsync(id, cancellationToken);
         }
         catch (Exception ex)
